client/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
Flask==1.1.2